Fix issue where calling Arg directly as a function no longer worked
