+++ .babelrc
@@ -1,0 +1,57 @@
+{
+    "presets": [
+        [
+            "@babel/preset-env",
+            {
+                "targets": {
+                    "browsers": [
+                        "> 2%",
+                        "last 3 versions",
+                        "not ie < 11",
+                        "not blackberry > 0",
+                        "not UCAndroid > 0",
+                        "not Samsung > 0",
+                        "not baidu > 0",
+                        "not QQAndroid > 0",
+                        "not Opera > 0",
+                        "not OperaMini all",
+                        "not OperaMobile > 0",
+                        "not Android > 0"
+                    ],
+                    "node": "current"
+                },
+                "useBuiltIns": "entry"
+            }
+        ],
+        "@babel/preset-react"
+    ],
+    "plugins": [
+        "@babel/plugin-transform-runtime",
+        "@babel/plugin-proposal-export-default-from",
+        "@babel/plugin-proposal-logical-assignment-operators",
+        "@babel/plugin-proposal-optional-chaining",
+        [
+            "@babel/plugin-proposal-pipeline-operator",
+            {
+                "proposal": "minimal"
+            }
+        ],
+        "@babel/plugin-proposal-nullish-coalescing-operator",
+        "@babel/plugin-proposal-do-expressions",
+        [
+            "@babel/plugin-proposal-decorators",
+            {
+                "legacy": true
+            }
+        ],
+        "@babel/plugin-proposal-function-sent",
+        "@babel/plugin-proposal-export-namespace-from",
+        "@babel/plugin-proposal-numeric-separator",
+        "@babel/plugin-proposal-throw-expressions",
+        "@babel/plugin-syntax-dynamic-import",
+        "@babel/plugin-syntax-import-meta",
+        "@babel/plugin-proposal-class-properties",
+        "@babel/plugin-proposal-json-strings",
+        "react-hot-loader/babel"
+    ]
+}
