+++ .editorconfig
@@ -1,0 +1,14 @@
+root = true
+
+[*.{js,jsx,scss,html}]
+charset = utf-8
+indent_style = space
+indent_size = 4
+end_of_line = lf
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+# https://github.com/npm/npm/pull/3180#issuecomment-16336516
+[{package.json}]
+indent_style = space
+indent_size = 2
