+++ .eslintrc
@@ -1,0 +1,13 @@
+{
+    "extends": "@spothero",
+    "overrides": [
+        {
+            "files": [
+                "**/*.spec.js"
+            ],
+            "rules": {
+                "max-nested-callbacks": 0
+            }
+        }
+    ]
+}
