+++ cypress.json
@@ -1,0 +1,4 @@
+{
+    "baseUrl": "http://localhost:3000",
+    "testFiles": "**/*.spec.js"
+}
