+++ index.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="x-ua-compatible" content="ie=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+
+        <title>Arg.js</title>
+    </head>
+    <body>
+        <div class="root"></div>
+
+        <!-- build:js -->
+        <script src="/js/bundle.js"></script>
+        <!-- endbuild -->
+    </body>
+</html>
