+++ js/main.js
@@ -1,0 +1,4 @@
+import Arg from '../src/arg';
+
+console.log(Arg.get('test'));
+console.log(Arg('test'));
